sw_apps :zynqmp_fsbl: Block addresses for higher DDR are not passed correctly to XFsbl_SetTlbAttributes function for A53 FSBL
The block addresses are of 64 bit and they are calculated using 32bit
variable called BlockNum. Changing variable type to u64 fixes the issue.

Acked-by: Krishna Chaitanya Patakamuri <kpataka@xilinx.com>

Acked-by: Krishna Chaitanya Patakamuri <kpataka@xilinx.com>

+++ lib/sw_apps/zynqmp_fsbl/src/xfsbl_initialization.c
@@ -620,7 +620,7 @@ static u32 XFsbl_SystemInit(XFsblPs * FsblInstancePtr)
 {
 	u32 Status;
 #if defined (XPAR_PSU_DDR_0_S_AXI_BASEADDR) && !defined (ARMR5)
-	u32 BlockNum;
+	u64 BlockNum;
 #endif
 
 	if (FsblInstancePtr->ResetReason != XFSBL_PS_ONLY_RESET) {
